requirements.txt
# This file may be used to create an environment using:
# $ conda create --name nlp --file <this file>
# platform: win-64
blas=1.0=mkl
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotli-python=1.0.9=py38hd77b12b_7
ca-certificates=2023.08.22=haa95532_0
certifi=2023.7.22=py38haa95532_0
cffi=1.15.1=py38h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=py38haa95532_0
contourpy=1.0.5=py38h59b6b97_0
cpuonly=2.0=0
cryptography=41.0.3=py38h89fc84f_0
cycler=0.11.0=pyhd3eb1b0_0
filelock=3.9.0=py38haa95532_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
giflib=5.2.1=h8cc25b3_3
icu=73.1=h6c2663c_0
idna=3.4=py38haa95532_0
importlib_resources=6.1.0=py38haa95532_0
intel-openmp=2023.1.0=h59b6b97_46319
jinja2=3.1.2=py38haa95532_0
jpeg=9e=h2bbff1b_1
kiwisolver=1.4.4=py38hd77b12b_0
krb5=1.20.1=h5b6d351_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_0
libjpeg-turbo=2.0.0=h196d8e1_0
libpng=1.6.39=h8cc25b3_0
libpq=12.15=h906ac69_1
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.1=py38h2bbff1b_0
matplotlib=3.7.2=py38haa95532_0
matplotlib-base=3.7.2=py38h4ed8f06_0
mkl=2023.1.0=h6b88ed4_46357
mkl-service=2.4.0=py38h2bbff1b_1
mkl_fft=1.3.8=py38h2bbff1b_0
mkl_random=1.2.4=py38h59b6b97_0
mpmath=1.3.0=py38haa95532_0
munkres=1.1.4=py_0
networkx=3.1=py38haa95532_0
numpy=1.24.3=py38h79a8e48_1
numpy-base=1.24.3=py38h8a87ada_1
openjpeg=2.4.0=h4fc8c34_0
openssl=3.0.12=h2bbff1b_0
packaging=23.1=py38haa95532_0
pillow=10.0.1=py38h045eedc_0
pip=23.3=py38haa95532_0
ply=3.11=py38_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py38haa95532_0
pyparsing=3.0.9=py38haa95532_0
pyqt=5.15.10=py38hd77b12b_0
pyqt5-sip=12.13.0=py38h2bbff1b_0
pysocks=1.7.1=py38haa95532_0
python=3.8.18=h1aa4202_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=2.1.0=py3.8_cpu_0
pytorch-mutex=1.0=cpu
pyyaml=6.0.1=py38h2bbff1b_0
qt-main=5.15.2=h19c9488_10
requests=2.31.0=py38haa95532_0
setuptools=68.0.0=py38haa95532_0
sip=6.7.12=py38hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h2bbff1b_0
sympy=1.11.1=py38haa95532_0
tbb=2021.8.0=h59b6b97_0
tk=8.6.12=h2bbff1b_0
tomli=2.0.1=py38haa95532_0
torchaudio=2.1.0=py38_cpu
torchvision=0.16.0=py38_cpu
tornado=6.3.3=py38h2bbff1b_0
tqdm=4.65.0=py38hd4e2768_0
typing_extensions=4.7.1=py38haa95532_0
urllib3=1.26.18=py38haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.41.2=py38haa95532_0
win_inet_pton=1.1.0=py38haa95532_0
xz=5.4.2=h8cc25b3_0
yaml=0.2.5=he774522_0
zipp=3.11.0=py38haa95532_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
